pyproject.toml
[project]
name = "systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "cryptography>=45.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "oss2>=2.19.1",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.45",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.4.6",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
